move MessagingDelegate to internal sub package
